pyproject.toml
[tool.poetry]
name = "monugstore"
version = "0.1.0"
description = "A quick python cookbook for dealing with google-cloud-storage"
authors = ["Pedro Cavalcanti"]
license = "MIT"
readme = "README.md"

# Add this section to include the scripts folder
include = [
    "scripts/*"
]

[tool.poetry.scripts]    
dump-key = "monugstore.scripts.dump_json:main"

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-storage = "^2.18.2"
pydantic = "^2.9.2"
google-cloud-secret-manager = "^2.20.2"
python-dotenv = "^1.0.1"
pymongo = "^4.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
